requirements.txt
pandas==2.0.3
numpy==1.24.3
plotly==5.15.0
matplotlib==3.7.1
seaborn==0.12.2
jupyter==1.0.0
kaleido==0.2.1
reportlab==4.0.4
markdown2==2.4.10
